+++ R/confint.R
@@ -15,7 +15,8 @@ function(object, parm, level = 0.95, ...) {
     a <- c(a, 1 - a)
     fac <- qnorm(a)
     ci <- array(NA, dim = c(length(parm), 2L), dimnames = list(parm, a))
-    ses <- sqrt(attributes(object)[["Variances"]])
+    variances <- attributes(object)[["Variances"]]
+    ses <- if (is.null(variances)) rep(NA_real_, length(cf)) else sqrt(variances)
     ci[] <- cf + ses %o% fac
     colnames(ci) <- sprintf("%0.2f%%", 100 * a)
     ci