+++ R/prediction.R
@@ -6,7 +6,7 @@
 #' @param type A character string indicating the type of marginal effects to estimate. Mostly relevant for non-linear models, where the reasonable options are \dQuote{response} (the default) or \dQuote{link} (i.e., on the scale of the linear predictor in a GLM).
 #' @param \dots Additional arguments passed to \code{\link[stats]{predict}} methods.
 #' @details This function is simply a wrapper around \code{\link[stats]{predict}} that returns a data.frame containing predicted values with respect to all variables specified in \code{data}. It is used internally by \code{\link{build_margins}}.
-#' @return An data.frame with number of rows equal to number of rows in \code{data}, where each row is an observation and the two columns represent fitted/predicted values and the standard errors thereof.
+#' @return A data.frame with class \dQuote{prediction} that has a number of rows equal to number of rows in \code{data}, where each row is an observation and the two columns represent fitted/predicted values (\code{fitted}) and the standard errors thereof (\code{se.fitted}).
 #' require("datasets")
 #' x <- lm(mpg ~ cyl * hp + wt, data = mtcars)
 #' prediction(x)
@@ -40,7 +40,7 @@ prediction.lm <- function(model, data, type = "response", ...) {
     # obs-x-2 data.frame of predictions
     structure(list(fitted = pred[["fit"]], 
                    se.fitted = pred[["se.fit"]]), 
-              class = "data.frame", 
+              class = c("prediction", "data.frame"), 
               row.names = seq_len(length(pred[["fit"]])))
 }
 
@@ -66,6 +66,31 @@ prediction.glm <- function(model, data, type = c("response", "link"), ...) {
     # obs-x-2 data.frame of predictions
     structure(list(fitted = pred[["fit"]], 
                    se.fitted = pred[["se.fit"]]), 
-              class = "data.frame", 
+              class = c("prediction", "data.frame"), 
+              row.names = seq_len(length(pred[["fit"]])))
+}
+#' @rdname prediction
+#' @export
+prediction.loess <- function(model, data, type = "response", ...) {
+    # setup data
+    if (missing(data)) {
+        if (!is.null(model[["call"]][["data"]])) {
+            data <- eval(model[["call"]][["data"]], parent.frame()) 
+        } else { 
+            data <- get_all_vars(model[["terms"]], data = model[["model"]])
+        }
+    }
+    
+    type <- match.arg(type)
+    
+    # extract predicted value at input value (value can only be 1 number)
+    pred <- predict(model, newdata = data, type = type, se = TRUE, ...)
+    class(pred[["fit"]]) <- c("fit", "numeric")
+    class(pred[["se.fit"]]) <- c("se.fit", "numeric")
+    
+    # obs-x-2 data.frame of predictions
+    structure(list(fitted = pred[["fit"]], 
+                   se.fitted = pred[["se.fit"]]), 
+              class = c("prediction", "data.frame"), 
               row.names = seq_len(length(pred[["fit"]])))
 }