+++ R/summary.R
@@ -12,9 +12,10 @@ function(object, digits = 4, level = 0.95, ...) {
     }
     fmt <- paste0("%0.", ifelse(digits > 7, 7, digits), "f")
     mes <- extract_marginal_effects(object)
+    variances <- attributes(mes)[["Variances"]]
     tab <- structure(list(Factor = names(mes), 
-                          "dy/dx" = colMeans(mes),
-                          "Std.Err." = sqrt(attributes(mes)[["Variances"]])
+                          "dy/dx" = colMeans(mes, na.rm = TRUE),
+                          "Std.Err." = if (is.null(variances)) rep(NA_real_, ncol(mes)) else sqrt(variances)
                           ),
                      class = "data.frame", row.names = names(mes))
     tab[["z value"]] <- tab[,"dy/dx"]/tab[,"Std.Err."]
@@ -29,5 +30,5 @@ function(object, digits = 4, level = 0.95, ...) {
 #' @export
 summary.marginslist <- 
 function(object, row.names = FALSE, ...) {
-    invisible(lapply(object, summary, ...))
+    lapply(object, summary, ...)
 }