+++ NEWS.md
@@ -3,6 +3,9 @@
 ## margins 0.2.7
 
 * `marginal_effects()` and `prediction()` are now S3 generics, with methods for "lm" and "glm" objects, improving extensability. (#39, #40)
+* Added preliminary support for "loess" objects, including methods for `prediction()`, `marginal_effects()`, `cplot()`, and `persp()`. No effect variances are currently calculated.
+* `prediction()` returns a new class ("prediction") and gains a `print()` method.
+* Internal function `get_effect_variances()` gains a "none" option for the `vce` argument, to skip calculation of ME variances.
 
 ## margins 0.2.7